Add dataB and recompute processInitial; fix plotting
Added a new feature to the workflow, triggered by this example.
o Re-run only the data15 and data16 with a different selection (veto events with badbatman jets).
o Copied the Reader to another folder, ran the initial step to merge them.
o Renamed the data15->data15B, same for data16, and copied in the folder where the current data15, data16 existed.
o Updated code to know about these new processes. I want to compare in the same tables, plots to see the impact.
o Updated function that calculates the list of available processInitial to be either as until know, from the fetch directory, or the new feature, from the processInitial. It then picked up the new process and plotted it.

Fixed plotting:
o there were issues on the edges of the eta plots.
o fixed bug in binRange calculation
o updated the workflow for plotting: first rebin to the non-uniform bins, then use the new feature of adding under/over-flows to the edge bins, and then rescale the histogram to get averages. The plots look nicer.